requirements.txt
streamlit
pandas
google-genai
requests
plotly
lxml
xlsxwriter
